adds delete button for failed tenants
