+++ webapp/src/main/webapp/src/app/admin/tenant/component/tenant-link/tenant-link.component.html
@@ -1,22 +1,32 @@
 <a
   href="javascript:void(0)"
   class="status-{{ tenant.status.toLowerCase() }}"
-  (click)="tenantClick.emit(tenant)"
+  (click)="$event.stopPropagation(); tenantClick.emit(tenant)"
 >
   <div class="head">
     <h2 class="h3">
       {{ tenant.label }}
       <span class="sandbox-key">({{ tenant.code }})</span>
     </h2>
-    <div
-      class="label border-only"
-      [ngClass]="{
-        'blue-dark': tenant.status === 'ACTIVE',
-        aqua: tenant.status !== 'ACTIVE' && tenant.status !== 'FAILED',
-        maroon: tenant.status === 'FAILED'
-      }"
-    >
-      {{ 'common.tenant-status.' + tenant.status | translate }}
+    <div class="menu">
+      <div
+        class="label border-only"
+        [ngClass]="{
+          'blue-dark': tenant.status === 'ACTIVE',
+          aqua: tenant.status !== 'ACTIVE' && tenant.status !== 'FAILED',
+          maroon: tenant.status === 'FAILED'
+        }"
+      >
+        {{ 'common.tenant-status.' + tenant.status | translate }}
+      </div>
+      <button
+        *ngIf="tenant.status === 'FAILED'"
+        type="button"
+        class="btn btn-danger btn-xs"
+        (click)="$event.stopPropagation(); tenantDelete.emit(tenant)"
+      >
+        {{ 'common.action.delete' | translate }}
+      </button>
     </div>
   </div>
   <div class="body">
