+++ webapp/src/main/webapp/src/app/admin/tenant/component/tenant-link/tenant-link.component.ts
@@ -19,4 +19,7 @@ export class TenantLinkComponent {
 
   @Output()
   tenantClick: EventEmitter<TenantConfiguration> = new EventEmitter();
+
+  @Output()
+  tenantDelete: EventEmitter<TenantConfiguration> = new EventEmitter();
 }
