+++ webapp/src/main/webapp/src/app/admin/tenant/pages/tenant/tenant.component.ts
@@ -12,19 +12,17 @@ import {
 import { TenantService } from '../../service/tenant.service';
 import { DataSet, TenantConfiguration } from '../../model/tenant-configuration';
 import { Observable, Subject } from 'rxjs';
-import { BsModalRef, BsModalService } from 'ngx-bootstrap';
 import { UserService } from '../../../../shared/security/service/user.service';
 import { LanguageStore } from '../../../../shared/i18n/language.store';
 import { NotificationService } from '../../../../shared/notification/notification.service';
 import { RdwTranslateLoader } from '../../../../shared/i18n/rdw-translate-loader';
-import { ConfirmationModalComponent } from '../../../../shared/component/confirmation-modal/confirmation-modal.component';
-import { TranslateService } from '@ngx-translate/core';
 import { of } from 'rxjs/internal/observable/of';
 import { FormMode } from '../../component/tenant-form/tenant-form.component';
 import { TenantType } from '../../model/tenant-type';
 import { combineLatest } from 'rxjs/internal/observable/combineLatest';
 import { defaultTenant } from '../../model/tenants';
 import { flatten } from '../../../../shared/support/support';
+import { TenantModalService } from '../../service/tenant-modal.service';
 
 @Component({
   selector: 'app-tenant',
@@ -50,10 +48,9 @@ export class TenantComponent implements OnDestroy {
     private service: TenantService,
     private userService: UserService,
     private languageStore: LanguageStore,
-    private modalService: BsModalService,
+    private modalService: TenantModalService,
     private notificationService: NotificationService,
-    private translationLoader: RdwTranslateLoader,
-    private translateService: TranslateService
+    private translationLoader: RdwTranslateLoader
   ) {
     const sandboxDataSets$ = this.service.getSandboxDataSets().pipe(share());
     const tenants$ = this.service.getAll('TENANT').pipe(share());
@@ -158,39 +155,7 @@ export class TenantComponent implements OnDestroy {
   }
 
   onDelete(tenant: TenantConfiguration): void {
-    const modalReference: BsModalRef = this.modalService.show(
-      ConfirmationModalComponent
-    );
-    const modal: ConfirmationModalComponent = modalReference.content;
-
-    modal.head = this.translateService.instant(
-      'tenant.delete.modal.head',
-      tenant
-    );
-    modal.body = this.translateService.instant(
-      'tenant.delete.modal.body',
-      tenant
-    );
-    modal.acceptButton = this.translateService.instant('common.action.delete');
-    modal.acceptButtonClass = 'btn-danger';
-    modal.declineButton = this.translateService.instant('common.action.cancel');
-    modal.accept.subscribe(() => {
-      this.service.delete(tenant.code).subscribe(
-        () => {
-          this.notificationService.success({
-            id: 'tenant.delete.success'
-          });
-          this.router.navigate(['..'], {
-            relativeTo: this.route
-          });
-        },
-        error => {
-          this.notificationService.error({
-            id: 'tenant.delete.error'
-          });
-        }
-      );
-    });
+    this.modalService.openDeleteConfirmationModal(tenant);
   }
 
   private onSave(mode: 'create' | 'update', value: TenantConfiguration): void {
