+++ webapp/src/main/webapp/src/styles.less
@@ -60,6 +60,11 @@
 /**
  * Bootstrap additions
  */
+.modal-footer {
+  .btn + .btn {
+    margin-left: 10px;
+  }
+}
 .dropdown-menu > li > a:hover,
 .dropdown-menu > li > a:focus {
   background-color: #f0f0f0; /* was #f8f8f8 */
